requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.5.6
attrs==21.1.0
bcrypt==3.2.0
black==21.5b0
blinker==1.4
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-restx==0.3.0
Flask-Script==2.0.6
Flask-Testing==0.8.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.23.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.2
pymongo==3.11.4
pyrsistent==0.17.3
pytz==2021.1
regex==2021.4.4
six==1.16.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
